api/requirements.txt
Flask==2.0.1
python-dotenv==0.19.2
Flask-PyMongo==2.3.0
flask-jwt-extended==4.5.2